pyproject.toml
[project]
name = "a2a-cli"
version = "0.1.0"
description = "A2A CLI Client and Server Application"
authors = [
    { name = "A2A CLI Developer", email = "dev@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.9",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "pydantic>=2.10.6",
    "uvicorn>=0.34.0",
    "asyncclick>=8.1.8",
    "rich>=13.0.0",
    "fastapi>=0.115.2",
    "starlette>=0.46.1",
    # LLM Provider dependencies
    "google-generativeai>=0.8.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "aiofiles>=24.1.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "mypy>=1.15.0",
    "ruff>=0.11.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
a2a-client = "client.main:main"
a2a-server = "server.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "mypy>=1.15.0",
    "ruff>=0.11.6",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "B", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true